Fixes timing issue during delete
- Waits for vpn server to be destroyed before returning control - closes #12
- Updates workflows to use ones in action-workflows - closes #13
- Updates clis submodule to v1.12.0 so that plugins are updated

+++ scripts/create-certificates.sh
@@ -1,4 +1,11 @@
-PATH=$BIN_DIR:$PATH
+#!/usr/bin/env bash
+
+export PATH=$BIN_DIR:$PATH
+
+if ! command -v git 1> /dev/null 2> /dev/null; then
+  echo "git cli not found" >&2
+  exit 1
+fi
 
 echo "Cloning easy-rsa"
 rm -rf easy-rsa
